Move settings into /data
