+++ scripts/setup.mjs
@@ -2,7 +2,7 @@ import {Crafting} from "./modules/data/crafting/base-crafting.mjs";
 import {Mayhem} from "./modules/data/mayhem.mjs";
 import {Resting} from "./modules/data/resting.mjs";
 import {PublicInterface} from "./modules/interface.mjs";
-import {Settings} from "./modules/settings.mjs";
+import {Settings} from "./modules/data/settings.mjs";
 import {SystemConfig} from "./modules/system-config.mjs";
 
 Hooks.once("init", SystemConfig.init);
